+++ lib/src/utils/types_def.dart
@@ -2,3 +2,4 @@ const type$id = r'$oid';
 const type$int32 = r'$numberInt';
 const type$int64 = r'$numberLong';
 const type$date = r'$date';
+const type$code = r'$code';